requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.87
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
etils==1.11.0
filelock==3.16.1
flax==0.10.2
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
jax==0.4.36
jax-cuda12-pjrt==0.4.36
jax-cuda12-plugin==0.4.36
jaxlib==0.4.36
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.10.2
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
pybullet==3.2.6
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
ray==2.40.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
simplejson==3.19.3
six==1.17.0
sympy==1.13.1
tensorstore==0.1.69
toolz==1.0.0
torch==2.5.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.21.0
